Update generator for Unicode 12.1 (#426)
* Update generator for latest emoji data and dependencies

This comes with some significant changes since the generated code as the
structure was before could not be compiled. We now chunk the emojis into
configurable lists which each get a class in the style of
ActivitiesCategoryChunk0. Also, the categories have changed a bit, so these
get now normalized to what we had before. Finally this fixes a bug causing
low(er) image quality.

* Add facebook module

* Fix code style issues

* Also generate strings resource

* Add google-compat to the generator

* Minor adjustments

* Add generated files